feat(rate-limiting) add support for draft-rfc rate-limiting headers
### Summary

This PR adds support for the following headers specified in
https://tools.ietf.org/html/draft-polli-ratelimit-headers-01:

- RateLimit-Limit
- RateLimit-Remaining
- RateLimit-Reset

For `RateLimit-Limit` this also implements multiple windows support:
https://tools.ietf.org/html/draft-polli-ratelimit-headers-01#section-6.2.4

(if reviewers think that we should remove it, let me know)

The multiple windows are enabled only when there is more than one
rate-limiting window configured for the plugin. Otherwise it just
sends a single integer on `RateLimit-Limit` header.

All this work was done by @ioggstream (give him a t-shirt or two!) on
several attempts: #5330 and #3451, and btw. @ioggstream is the author
of that draft rfc, which I hope gets promoted to a proposed standard!
Great stuff! And sure we want to support it.

I (@bungle) gathered the ideas and feedback from the previous PRs
(including #2735) and tried to make all these in a state where we
can agree and merge. I have refactored the code a bit when adding
these new headers. And I also added a bunch of tests.

This PR does not touch the previous `properietary` headers the plugin
has sent for a long time, namely:
- X-RateLimit-Limit-*
- X-RateLimit-Remaining-*

We can later on decide, perhaps when this is standardized (or some time
after that), if we want to drop support on these, but at this point
they remain.