feat(db) make path_handling setting impervious to migration errors
The previous way of detecting path_handling setting had an important problem: it could be confused by failed migrations.

Say we're upgrading from `001_14_to_15`.

When we are on the `up` phase of the `007`, we detect this, and set the path_handling of any existing route to v0.

The migrations continue, starting the `teardown` phase now. An error gets raised on the `004_110_to_120` migration. This means that the "last run migration" will be reset in the database to `003_100_to_110`.

The user now runs `kong migrate` again. Since the last migration run is now `003_...` instead of `001_...`, we will try to set the path handling to `v1` instead of `v0`.

This change attempts to fix this is by marking the version we should use on the database, the first time we see it. If we already have written down a path_handling in the database, we use that. Otherwise we calculate it and we store it in the database.

We reuse the `schema_meta` table with a new key so that it doesn't conflict with other schema_meta properties.